refactor: move actual workings to their own module
